Optimize performance
Use FX2D renderer which blows away default and faster than P2D and P3D.

Set loadPercentage to 1.0 so a particle is created for each pixel.

Set particleSize to 1 for closer correspondence to pixels. Actual
strokeWeight is interpolated but this keeps it approximately 1. Larger
weights can have a major performance impact.